% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateRegions.R
\name{annotateRegions}
\alias{annotateRegions}
\title{annotateRegions}
\usage{
annotateRegions(
  regions,
  anno,
  proximal = c(2500, 1000),
  filter = AnnotationFilterList(),
  extra = list(),
  ignore.strand = TRUE,
  ...
)
}
\arguments{
\item{regions}{A GRanges object}

\item{anno}{An annotation object, such as an \code{\link[ensembldb]{EnsDb}})
object, a TxDb object, or a GRanges object of a GENCODE-like gtf or the 
path to such a file.}

\item{proximal}{The threshold(s) for TSS proximal regions. Multiple values
will result in multiple class factor levels.}

\item{filter}{An \code{\link[AnnotationFilter]{AnnotationFilter}} to filter
transcripts. Only used if `anno` is an \code{\link[ensembldb]{EnsDb}}).}

\item{extra}{An optional named list of GRanges for additional overlaps. Each
list element will create an additional binary metadata column.}

\item{ignore.strand}{Whether to ignore the strand for the overlap with 
elements of `extra` (default TRUE).}

\item{...}{Passed to \code{\link[GenomicRanges]{overlapsAny}} for the 
overlaps with `extra`.}
}
\value{
The sorted `regions` object with additional annotation columns.
}
\description{
Annotates a GRanges on the basis of an annotation object (e.g. 
\code{\link[ensembldb]{EnsDb}}).
}
\examples{
# we first create some regions we want to annotate:
regions <- as(c("chrY:2655742-2655890", "chrY:28730110-28730950"), "GRanges")
# we'll use a lightweight ensembldb annotation for the annotation:
chrY <- system.file("chrY", package="ensembldb")
edb <- EnsDb(makeEnsemblSQLiteFromTables(path=chrY ,dbname=tempfile()))
# we run teh annotation:
regions <- annotateRegions(regions, edb)
# this adds metadata columns to the regions:
regions
table(regions$class)
}
