% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bam2bw.R
\name{bam2bw}
\alias{bam2bw}
\title{bam2bw}
\usage{
bam2bw(
  bamfile,
  output_bw,
  paired = NULL,
  binWidth = 20L,
  extend = 0L,
  scaling = TRUE,
  type = c("full", "center", "start", "end"),
  strand = c("*", "+", "-"),
  filter = 1L,
  shift = 0L,
  log1p = FALSE,
  includeDuplicates = TRUE,
  includeSecondary = TRUE,
  minMapq = 1L,
  minFragLength = 1L,
  maxFragLength = 5000L,
  splitByChr = 3,
  ...
)
}
\arguments{
\item{bamfile}{The path to the bam file.}

\item{output_bw}{The path to the output bigwig file}

\item{paired}{Logical; whether to consider fragments instead of reads for 
paired-end data. For spliced (e.g. RNA-seq) data, paired should always be 
set to FALSE.}

\item{binWidth}{The window size. A lower value (min 1) means a higher 
resolution, but larger file size.}

\item{extend}{How much to extend single-end reads}

\item{scaling}{Either TRUE (performs Count Per Million scaling), FALSE (no 
scaling), or a numeric value by which the signal will be divided.}

\item{type}{Type of the coverage to compile. Either full (full read/fragment),
start (count read/fragment start locations), end, or center.}

\item{strand}{Strand(s) to capture}

\item{filter}{Minimum count for a bin to be considered.}

\item{shift}{Shift (from 3' to 5') by which reads/fragments will be shifted.}

\item{log1p}{Whether to log-transform (`log(x+1)`) the (scaled) signal.}

\item{includeDuplicates}{Logical, whether to include reads flagged as 
duplicates.}

\item{includeSecondary}{Logical; whether to include secondary alignments}

\item{minMapq}{Minimum mapping quality (1 to 255)}

\item{minFragLength}{Minimum fragment length (ignored if `paired=FALSE`)}

\item{maxFragLength}{Maximum fragment length (ignored if `paired=FALSE`)}

\item{splitByChr}{Whether to process chromosomes separately, and if so by how
many chunks. The should not affect the output, and is simply slightly 
slower and consumes less memory. Can be a logical value, but we instead
recommend giving a positive integer indicating the number of chunks.}

\item{...}{Passed to `ScanBamParam`}
}
\value{
The bigwig filepath
}
\description{
Creates a coverage bigwig file from a bam file.
The implementation requires loading the reads in memory, which can be quite
memory-hungry, so consider setting `splitByChr` to an integer above 1 to split
the work into chunks.
}
\examples{
# get an example bam file
bam <- system.file("extdata", "ex1.bam", package="Rsamtools")
# create bigwig
bam2bw(bam, scaling=TRUE)
}
