% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bamChrChunkApply.R
\name{bamChrChunkApply}
\alias{bamChrChunkApply}
\title{bamChrChunkApply}
\usage{
bamChrChunkApply(
  x,
  FUN,
  paired = FALSE,
  keepSeqLvls = NULL,
  nChunks = 4,
  strandMode = 2,
  flgs = scanBamFlag(),
  exclude = NULL,
  mapqFilter = NA_integer_,
  BPPARAM = SerialParam(),
  ...
)
}
\arguments{
\item{x}{A bam file.}

\item{FUN}{The function to be run, the first argument of which should be a
`GRanges`}

\item{paired}{Logical; whether to consider the reads as paired (fragments, 
rather than reads, will be returned)}

\item{keepSeqLvls}{An optional vector of seqLevels to keep}

\item{nChunks}{The number of chunks to use (higher will use less memory but
increase overhead)}

\item{strandMode}{Strandmode for paired data (see 
\code{\link[GenomicAlignments]{readGAlignmentPairs}}).}

\item{flgs}{`scanBamFlag` for filtering the reads}

\item{exclude}{An optional GRanges of regions for which overlapping reads 
should be excluded.}

\item{mapqFilter}{Integer of the minimum mapping quality for reads to be 
included.}

\item{BPPARAM}{A `BiocParallel` parameter object for multithreading. Note 
that if used, memory usage will be high; in this context we recommend a 
high `nChunks`.}

\item{...}{Passed to `FUN`}
}
\value{
A list of whatever `FUN` returns
}
\description{
Runs a function on reads/fragments from chunks of (chromosomes) of an 
indexed bam file. This is especially used by other functions to avoid 
loading all alignments into memory, or to parallelize reads processing.
}
