% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{breakStrings}
\alias{breakStrings}
\title{breakStrings}
\usage{
breakStrings(x, minSizeForBreak = 20, lb = "\\n")
}
\arguments{
\item{x}{A character (or factor) vector.}

\item{minSizeForBreak}{The minimum number of characters to be broken into 
two lines.}

\item{lb}{The separation character (e.g. line break).}
}
\value{
A character vector of length=length(x).
}
\description{
breaks a string of long-enough words (or vector thereof) into two lines.
}
\examples{
breakStrings("this is too long for practical purposes")
}
