% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{bwNormFactors}
\alias{bwNormFactors}
\title{bwNormFactors}
\usage{
bwNormFactors(x, ...)
}
\arguments{
\item{x}{A vector of paths to bigwig files.}

\item{...}{Passed to \code{\link{getNormFactors}}}
}
\value{
A vector of normalization factors
}
\description{
This function is deprecated, see \code{\link{getNormFactors}}.
}
