% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callPeaks.R
\name{callPeaks}
\alias{callPeaks}
\title{callPeaks}
\usage{
callPeaks(
  bam,
  ctrl = NULL,
  paired,
  type = c("narrow", "broad"),
  fragLength = 200L,
  globalNullH = FALSE,
  gsize = NULL,
  blacklist = NULL,
  binSize = 10L,
  nChunks = 8L,
  flags = scanBamFlag(isDuplicate = FALSE),
  minPeakCount = 5L,
  minFoldEnr = 1.3,
  pthres = 10^-5,
  maxSize = NULL,
  bgWindow = c(1, 5, 10) * 1000,
  pseudoCount = 0.5,
  useStrand = !paired,
  outFormat = c("narrowPeak", "custom"),
  verbose = TRUE,
  ...
)
}
\arguments{
\item{bam}{A signal bam file (which should be accompanied by an index file).
Alternatively, a TABIX-indexed fragment file, or an RleList object.}

\item{ctrl}{An optional path to a control bam file. Alternatively, an 
RleList object.}

\item{paired}{Logical, whether the reads are paired.}

\item{type}{The type of peaks to identify ('narrow' or 'broad').}

\item{fragLength}{Fragment length. Ignored if `paired=TRUE`. If 
`useStrand=TRUE` (default), this is only used for the initial candidate 
region identification, and sizes are adjusted after, so it doesn't need to 
be very precise.}

\item{globalNullH}{Logical; whether to use a global expectation, rather than
the local background (default), in the absence of a control.}

\item{gsize}{The mappable genome size. Ignored unless `globalNullH=TRUE`.
Can also be a species acronym in 'hs', 'mm', 'dm', and 'ce'.}

\item{blacklist}{An optional `GRanges` of regions to be excluded (or the path
to such a file). Since the blacklisted regions are removed from both the
signal and control peaks, this also has an important impact on the 
empirical FDR (when `ctrl` is given).}

\item{binSize}{Binsize used to estimate peak shift.}

\item{nChunks}{The number of processing chunks. Increasing this will reduce
memory usage.}

\item{flags}{An optional \link[Rsamtools]{scanBamFlag} object to filter the 
reads. By default, reads flagged as optical duplicates are excluded.}

\item{minPeakCount}{The minimum summit count for a region to be considered.
Decreasing this can substantially increase the running time.}

\item{minFoldEnr}{The minimum fold-enrichment for a region to be 
considered. Decreasing this can substantially increase the running time.}

\item{pthres}{The p-value threshold to use.}

\item{maxSize}{The loose maximum size of a peak. This is the size above which
the method will attempt to break up peaks into smaller ones. By default, it
is 1000 for `type="narrow"`, and 5000 for `type="broad"`.}

\item{bgWindow}{The windows to consider (in addition to the peak itself) 
for local background.}

\item{pseudoCount}{The pseudocount to use when computing logFC.}

\item{useStrand}{Logical; whether to use strand information to better 
estimate the peak boundaries with single-end data.}

\item{outFormat}{The output format ('custom' or 'narrowPeak')}

\item{verbose}{Logical; whether to output progress messages}

\item{...}{Passed to \code{\link{bamChrChunkApply}}}
}
\value{
A `GRanges`, by default following the narrowPeak format 
  specification.
}
\description{
This is a native R peak caller loosely based on the general MACS2/3 strategy 
(Zhang et al., Genome Biology 2008). It can work from BAM files or from 
coverage tracks. The results are highly concordant with MACS, although the 
significance estimates are slightly more conservative and the function 
requires much more memory (see details).
}
\details{
Unless `globalNullH=TRUE`, this function uses MACS' local lambda (defined by
`bgWindow`). A major difference with MACS2/3 is that, rather than using 
sliding windows, if works on the running list encoding of the coverage(s).
As a consequence, significance is estimated based on the peak's maximum 
coverage, which is very similar for narrow peaks, but very different for 
broad peaks, which will not produce astronomical p-values as is the case 
with MACS. In the absence of a control, the peaks called with this function 
are slightly more reproducible across biological replicates. With a control, 
however, the 

Because FDR calculation is so tricky for peak calling, the function uses an 
uncorrected p-value threshold. If desired, users can further filter based on
the FDR. For more powerful FDR, we recommend the use of `blacklist`.

On a single thread, the function is nearly as fast as MACS2/3, but uses much 
more memory, chiefly due to the fact that reads are read into memory in 
large chunks. On a single thread, memory usage will be roughly the 
sum of the filesize of your bam files (i.e. IP + input, if using an input) 
divided by `nChunks-1`.

The function uses \code{\link{bamChrChunkApply}} to obtain the coverages,
and can accept any argument of that function. This means for instance that 
the `mapqFilter` argument can be used to restrict the reads used, or a 
`BPPARAM` argument to enable multi-threading (beware of memory consumption!).

To export as a narrowPeak file, see \code{\link{exportNarrowPeaks}}.
}
\examples{
# we use the example bam file from the Rsamtools package:
bam <- system.file("extdata", "ex1.bam", package="Rsamtools")
peaks <- callPeaks(bam, paired=TRUE)
}
