% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signalMatrices.R
\name{clusterSignalMatrices}
\alias{clusterSignalMatrices}
\title{clusterSignalMatrices}
\usage{
clusterSignalMatrices(
  ml,
  k,
  scaleRows = FALSE,
  scaleCols = FALSE,
  use = c("enrich", "full", "max", "center"),
  trim = c(0.05, 0.95),
  nstart = 3,
  ...
)
}
\arguments{
\item{ml}{A list of signal matrices, as produced by \code{\link{signal2Matrix}}.}

\item{k}{The number of clusters to generate}

\item{scaleRows}{Logical; whether to scale rows for clustering}

\item{scaleCols}{Logical; whether to scale columns (i.e. signals/samples)}

\item{use}{What values to use for clustering. By default, uses 
\code{\link[EnrichedHeatmap]{enriched_score}}. Other options are 'full' 
(uses the full signal for clustering), 'max' (uses the maximum value in 
the region), or 'center' (use the value at the center of the region).}

\item{trim}{Values to trim (applied individidually for each signal matrix)}

\item{nstart}{Number of starts for k-means clustering}

\item{...}{Passed to `kmeans`}
}
\value{
If `k` is of length 1, a vector of cluster labels, corresponding to 
  the rows of `ml`. If `length(k)>1`, a list of two data.frames containing:
  1) the cluster labels at the different resolutions, and 2) the variance 
  explained by clusters at each resolution.
}
\description{
clusterSignalMatrices
}
