% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlaps.R
\name{colOverlaps}
\alias{colOverlaps}
\title{colOverlaps}
\usage{
colOverlaps(x, y = NULL, metric = c("overlap", "jaccard", "overlapCoef"))
}
\arguments{
\item{x}{A logical matrix.}

\item{y}{An optional nother logical matrix or vector. If NULL (default), 
overlaps are computed between columns of `x`.}

\item{metric}{Either 'overlap', 'jaccard', or 'overlapCoef'.}
}
\value{
A matrix of pairwise metric values.
}
\description{
Computes pairwise overlap metric between columns
}
\examples{
m <- matrix(sample(c(TRUE,FALSE),12,replace=TRUE), nrow=4)
colOverlaps(m, metric="jaccard")
}
