% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifs.R
\name{discoverMotifInteractions}
\alias{discoverMotifInteractions}
\title{discoverMotifInteractions}
\usage{
discoverMotifInteractions(
  dev,
  group,
  covar = c(),
  global = FALSE,
  weights = c("fixed", "trend", "none"),
  useAssay = c("adjZ", "deviations", "z")
)
}
\arguments{
\item{dev}{A deviations SummarizedExperiment produced by 
\code{\link{getInteractionsDeviations}}}

\item{group}{The column of `colData(dev)` containing the experimental groups
of interest. To column will be coerced to a factor, and all comparisons 
will be made relative to the first level of the factor.}

\item{covar}{Optional columns of `colData(dev)` containing covariates to 
correct for.}

\item{global}{Logical; whether to test the contrasts globally, rather than
individually. Has no effect when there is a single contrast 
(i.e. two-group comparison).}

\item{weights}{The weights to use, either 'fixed' (uses the sqrt number of 
sites), 'trend' (default; uses the trend over the number of sites), or 
'none' (no weights; not recommended). We recommend 'trend' if the number
of motifs is sufficiently large (e.g. >100).}

\item{useAssay}{Which assay to use. We recommend setting 'adjZ' (default),
which uses z-scores but corrects the scale of the bait's z-scores to its
expectation for the intersection size.}
}
\value{
A data.frame.
}
\description{
Discover motifs interacting with a bait motif, based on the betterChromVAR
package.
}
\author{
Pierre-Luc Germain
}
