% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESE.R
\docType{class}
\name{EnrichmentSE-class}
\alias{EnrichmentSE-class}
\alias{.EnrichmentSE}
\alias{EnrichmentSE}
\alias{[,EnrichmentSE,ANY,ANY,ANY-method}
\alias{show,EnrichmentSE-method}
\alias{score,EnrichmentSE-method}
\title{EnrichmentSE class and constructor}
\usage{
EnrichmentSE(assays, rowRanges = NULL, ...)

\S4method{[}{EnrichmentSE,ANY,ANY,ANY}(x, i, j, ..., drop = TRUE)

\S4method{show}{EnrichmentSE}(object)

\S4method{score}{EnrichmentSE}(x, ...)
}
\arguments{
\item{assays}{A list of matrices or 
\code{\link[EnrichedHeatmap]{normalizedMatrix}} objects.}

\item{rowRanges}{A \code{\link[GenomicRanges]{GRanges}} object.}

\item{...}{Arguments passed to \code{\link{getSignalMatrices}}.}

\item{x}{An object of class `EnrichmentSE`, as produced by 
\code{\link{signal2Matrix}}.}

\item{i, j}{Indices for subsetting.}

\item{drop}{Logical; whether to drop dimensions.}

\item{object}{An `EnrichmentSE` object.}
}
\value{
An `EnrichmentSE` object.
}
\description{
The `EnrichmentSE` class is a container for epigenomic enrichment data, 
extending the \code{\link[SummarizedExperiment]{RangedSummarizedExperiment}}
 class.
}
\section{Functions}{
\itemize{
\item \code{x[i}: Subset method for EnrichmentSE

\item \code{show(EnrichmentSE)}: Show method for EnrichmentSE

\item \code{score(EnrichmentSE)}: Access the score (first assay)

}}
