% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateFragSize.R
\name{estimateFragSize}
\alias{estimateFragSize}
\title{estimateFragSize}
\usage{
estimateFragSize(
  bam,
  ctrl = NULL,
  binSize = 10L,
  mfold = c(10, 50),
  ...,
  minSummitCount = 8L,
  useSeqLevels = NULL,
  maxSize = 2500L,
  priorLength = 200L,
  blacklist = NULL,
  ret = c("mode", "median", "mean", "tables", "plots"),
  BPPARAM = SerialParam()
)
}
\arguments{
\item{bam}{The path to one or more bam files}

\item{ctrl}{Optional path to a control bam file (if `length(bam)>1`, the same
control will be used for all).}

\item{binSize}{Bin size. The precision of the reported fragment size is 
necessary lower than this. A higher bin size will improve the summit 
identification in low-coverage regions. We recommend leaving the default
value.}

\item{mfold}{The range of fold-enrichment over the control (if `ctrl` 
provided) or of coverages for the identification of regions based on which
distance will be estimated.}

\item{minSummitCount}{The minimum read count for a summit to be considered.}

\item{useSeqLevels}{An optional vector of seqLevels in which to conduct the
analysis.}

\item{maxSize}{The maximum size of regions to be used}

\item{priorLength}{The prior fragment length (use for read extension to 
identify enriched regions)}

\item{blacklist}{Optional `GRanges` of blacklisted regions to be excluded.}

\item{ret}{The type of return, either a 'table' of pairs of summits and their
properties, or a 'plot', or the median/mean/mode of the distance distribution.}

\item{BPPARAM}{A `BiocParallel` parameter object for multithreading. Only 
used if multiple files are given in `bam`.}
}
\value{
By default, the estimated (mode) fragment length(s), but see the 
  `ret` argument
}
\description{
estimateFragSize
}
\examples{
# get an example bam file
bam <- system.file("extdata", "ex1.bam", package="Rsamtools")
# create bigwig
estimateFragSize(bam)
}
