% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{exampleDNAme}
\alias{exampleDNAme}
\alias{geneBodies}
\title{Example DNAme data on some active gene bodies}
\value{
a named character vector of length 2.
}
\description{
A GRanges object (called `geneBodies`) containing the coordinates of some 
gene bodies on chr9 of hg38, as well as a GPos object (called `exampleDNAme`)
containing methylation percentages at CpGs in those regions. Taken from WG 
bisulfite sequencing data from A549, from ENCODE accession ID ENCFF948WVD.
}
