% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{exampleESE}
\alias{exampleESE}
\title{Example EnrichmentSE object}
\value{
a named character vector of length 1.
}
\description{
Small sample signal from ENCODE ChIP-seq for H3K27ac, H3K4me3 and p300, 
around some p300 binding sites and TSS in mESC.
}
