% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifs.R
\name{exploreMotifInteraction}
\alias{exploreMotifInteraction}
\title{exploreMotifInteraction}
\usage{
exploreMotifInteraction(
  counts,
  motifs,
  genome,
  bg = NULL,
  nDistQuantiles = 5L,
  maxDist = 500L,
  ...
)
}
\arguments{
\item{counts}{An object inheriting from `RangedSummarizedExperiment`, with
a `bias` column in its `rowData`.}

\item{motifs}{A named pair of motifs, in a format recognized by 
\code{\link[motifmatchr]{matchMotif}}.}

\item{genome}{A BSgenome or FaFile object.}

\item{nDistQuantiles}{The number of distance quantiles between the motif 
matches (up to `maxDist`)}

\item{maxDist}{The maximum distance between matches.}

\item{...}{Any further argument passed to 
\code{\link[betterChromVAR]{betterChromVAR}}.}
}
\value{
A SummarizedExperiment object of deviations for different types of
  interactions between the motifs.
}
\description{
We compare deviations in sites that harbor a single of the two motifs, or
the pair of motifs at given distances from each other.
}
\details{
Motif pairs are split according to the distance between their centers up to
`maxDist` (default 300) into `nDistQuantiles` (default 4) similarly-populated
quantiles. Deviations are reported for these different bins, as well as sites
harboring only a single motif.
}
\author{
Pierre-Luc Germain
}
