% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifs.R
\name{findMotifInstances}
\alias{findMotifInstances}
\title{findMotifInstances}
\usage{
findMotifInstances(seqs, motif, keepMatchedSeq = FALSE, ...)
}
\arguments{
\item{seqs}{A set of sequences, e.g. `DNAStringSet`, optionally with 
coordinate as names to enable conversion.}

\item{motif}{A motif, in any format recognized by `universalmotif`}

\item{keepMatchedSeq}{Logical; whether to keep the matched sequence.}

\item{...}{Passed to `TFBSTools::searchSeq`; can for instance be used to set
the number of threads to use, e.g. with `mc.cores=2`}
}
\value{
A `GRanges` object
}
\description{
A wrapper around `TFBSTools` for scanning motif occurence, handling the 
coordinate conversion as `memes` does. This is an alternative to the `memes`
package, which is more efficient but requires the meme suite to be available.
}
