% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{formatGenomicDist}
\alias{formatGenomicDist}
\title{formatGenomicDist}
\usage{
formatGenomicDist(
  e,
  allowFraction = TRUE,
  sameUnits = TRUE,
  head0 = TRUE,
  units = c(mb = 1e+06, kb = 1000, bp = 1)
)
}
\arguments{
\item{e}{An integer vector of genomic sizes/distances to be formatted}

\item{allowFraction}{Whether to allow decimals; can be either logical or a
number of decimals allowed (defaults to TRUE / 1)}

\item{sameUnits}{Logical; whether all values should get the same unit.}

\item{head0}{Logical; whether to keep heading zero}

\item{units}{The units to use.}
}
\value{
A character vector.
}
\description{
formatGenomicDist
}
