% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{fragSizesDist}
\alias{fragSizesDist}
\title{fragSizesDist}
\usage{
fragSizesDist(
  x,
  what = 10000,
  flags = scanBamFlag(isProperPair = TRUE),
  BPPARAM = SerialParam(),
  returnPlot = TRUE
)
}
\arguments{
\item{x}{A (named) vector of paths to bam files.}

\item{what}{Either a positive integer (length 1) indicating how many reads 
to randomly sample, or a character vector (of length 1) indicating which
chromosome to read.}

\item{flags}{A `scanBamFlag` object (see \link[Rsamtools]{ScanBamParam})}

\item{BPPARAM}{A \link[BiocParallel]{BiocParallel} BPPARAM object for
multithreading.}

\item{returnPlot}{Logical; whether to return a plot.}
}
\value{
If `returnPlot=TRUE`, returns a ggplot object, otherwise a
data.frame of fragment lengths.
}
\description{
fragSizesDist
}
\examples{
# example bam file:
bam <- system.file("extdata", "ex1.bam", package="Rsamtools")
fragSizesDist(bam, what=100)
}
