% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callPeaks.R
\name{getEmpiricalFDR}
\alias{getEmpiricalFDR}
\title{getEmpiricalFDR}
\usage{
getEmpiricalFDR(log10p, pneg, n = length(log10p) * 10)
}
\arguments{
\item{log10p}{-log10 p-values of candidate peaks}

\item{pneg}{-log10 p-values of negative peaks}

\item{n}{The number of hypotheses (used when the empirical FDR is zero)}
}
\value{
A data.frame with the empirical FDR and a smoothed -log10(FDR)
}
\description{
Computes the FDR as the proportion of negative peaks with a more extreme 
  p-value, eventually extrapolating and preserving the ranking.
}
