% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESE.R
\name{getSignalMatrices}
\alias{getSignalMatrices}
\title{getSignalMatrices}
\usage{
getSignalMatrices(x, assay = 1L)
}
\arguments{
\item{x}{An object of class `EnrichmentSE`, as produced by 
\code{\link{signal2Matrix}}.}

\item{assay}{The assay to extract (defaults to the first assay).}
}
\value{
A list of normalizedMatrix objects.
}
\description{
Extracts a list of signal matrices from an EnrichmentSE object.
}
\examples{
# we first create an EnrichmentSE object:
bw <- system.file("extdata/example_atac.bw", package="epiwraps")
regions <- rtracklayer::import(system.file("extdata/example_peaks.bed", 
                                           package="epiwraps"))
x <- signal2Matrix(bw, regions)
# then we can extract the list of signal matrices:
sm <- getSignalMatrices(x)
}
