% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{inject}
\alias{inject}
\title{Inject (insert) values at positions in a vector}
\usage{
inject(what, inWhat, at)
}
\arguments{
\item{what}{The values to inject}

\item{inWhat}{The vector in which to inject them}

\item{at}{The positions in the vector *after* which to inject the values.}
}
\value{
A vector of same mode at `inWhat`, and of length equal to the sum of 
  the lengths of `what` and `inWhat`.
}
\description{
Inject (insert) values at positions in a vector
}
\examples{
inWhat <- 1:10
inject(c(21,22,23), inWhat, at=as.integer(c(0,5,10)))
}
