% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signalMatrices.R
\name{mergeSignalMatrices}
\alias{mergeSignalMatrices}
\title{mergeSignalMatrices}
\usage{
mergeSignalMatrices(ml, aggregation = c("mean", "sum", "median"))
}
\arguments{
\item{ml}{A named list of matrices as produced by \code{\link{signal2Matrix}}}

\item{aggregation}{The method to aggregate matrices}
}
\value{
A `normalizedMatrix` object
}
\description{
mergeSignalMatrices
}
