% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifs.R
\name{motifFootprint}
\alias{motifFootprint}
\title{motifFootprint}
\usage{
motifFootprint(bamfile, motif, motif_occurences, genome = NULL, around = 100)
}
\arguments{
\item{bamfile}{The path to a bam file}

\item{motif}{The motif (as probability matrix)}

\item{motif_occurences}{A GRanges of the motif occurences}

\item{genome}{Optional genome used for seqlengths (otherwise estimated from
the occurences)}

\item{around}{How much around the motif to plot}
}
\value{
A plot
}
\description{
This is a wrapper around `ATACseqQC::factorFootprints`, making it compatible
with a broader variety of inputs and use cases.
}
