% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCounts.R
\name{peakPbCountsSE}
\alias{peakPbCountsSE}
\title{peakPbCountsSE}
\usage{
peakPbCountsSE(fragfile, peaks, bcmap, genome = NULL)
}
\arguments{
\item{fragfile}{The path to a Tabix-indexed fragment file.}

\item{peaks}{A GRanges of the regions in which to count.}

\item{bcmap}{A named vector, indicating the pseudobulk sample (values) in 
which to include each barcode (names).}

\item{genome}{A optional genome object or path to a genom fasta file. If
included, GC bias will be added to the rowData of the output object.}
}
\value{
A \link[SummarizedExperiment]{RangedSummarizedExperiment} with a 
  'counts' assay, and columns corresponding to each unique value of `bcmap`.
}
\description{
Generate a pseudobulk peak counts SummarizedExperiment.
}
