% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{plotCorFromCovStats}
\alias{plotCorFromCovStats}
\title{plotCorFromCovStats}
\usage{
plotCorFromCovStats(
  qc,
  method = c("pearson", "spearman"),
  col = NULL,
  na_col = "white",
  column_title = NULL,
  ...
)
}
\arguments{
\item{qc}{A list of coverage statistics, as produced by 
\code{\link{getCovStats}}.}

\item{method}{The correlation metrics to include}

\item{col}{Optional heatmap colors}

\item{na_col}{Color for the diagonal; passed to 
\code{\link[ComplexHeatmap]{Heatmap}}.}

\item{column_title}{Column title (if NULL, uses the metric)}

\item{...}{Passed to \code{\link[ComplexHeatmap]{Heatmap}}.}
}
\value{
A `Heatmap` or `HeatmapList` object ready to be plotted.
}
\description{
plotCorFromCovStats
}
