% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bam2bw.R
\name{reduceRleLists}
\alias{reduceRleLists}
\title{reduceRleLists}
\usage{
reduceRleLists(res, fn = "+")
}
\arguments{
\item{res}{A list of RleList objects}

\item{fn}{The function to use to combine them}
}
\value{
A combined RleList object.
}
\description{
reduceRleLists
}
\examples{
list_of_rlelists <- list(
  RleList(A=c(0,0,1,0,0), B=c(0,0,0,0,1)),
  RleList(A=c(1,1,1,0,0), C=c(0,1,1,1,1)) )
reduceRleLists(list_of_rlelists)
}
