% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refragment.R
\name{refragment}
\alias{refragment}
\title{refragment}
\usage{
refragment(
  bam,
  minSize = 20L,
  nfr = 120L,
  nuc = c(145, 190),
  binSize = 10L,
  verbose = TRUE,
  BPPARAM = BiocParallel::SerialParam()
)
}
\arguments{
\item{bam}{Path to a paired-end bam file, or a `GRanges` of fragments.}

\item{minSize}{Minimum fragment size}

\item{nfr}{Maximum size of nucleosome-free fragments}

\item{nuc}{Range (min & max) of mono-nucleosome fragment sizes}

\item{binSize}{Bin size for cut profile}

\item{verbose}{Logical; whether to print progress messages}

\item{BPPARAM}{`BiocParallel` param for multithreading.}
}
\value{
A list of processed fragment `GRanges`, including nucleosome-free,
  mono-nucleosome, and ambiguous fragments.
}
\description{
Uses the cut profiles to re-fragment in silico fragments that contain more 
  than a nucleosome.
}
