% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlaps.R
\name{regionOverlaps}
\alias{regionOverlaps}
\title{regionOverlaps}
\usage{
regionOverlaps(
  listOfRegions,
  mode = c("reduced", "pairwise"),
  ignore.strand = TRUE,
  cluster = length(listOfRegions) > 2,
  colorBy = c("overlapCoef", "jaccard"),
  ...,
  returnValues = FALSE,
  color = viridis::plasma(100),
  number_color = "black"
)
}
\arguments{
\item{listOfRegions}{A named list of two or more (non-empty) `GRanges`}

\item{mode}{Either 'reduced' or 'pairwise'. 'reduced' first uses `reduce` to
get a set of reference regions which are, based on overlap, contained or 
not in the different sets. It is thus symmetrical. `pairwise` does pairwise
overlap between the sets of regions; it is asymmetrical and slower to 
compute.}

\item{ignore.strand}{Logical; whether to ignore strand for overlaps (default
TRUE).}

\item{cluster}{Logical; whether to cluster rows/columns}

\item{colorBy}{Whether to color by 'overlapCoef' (default), or by 'jaccard'
index.}

\item{...}{Passed to \code{\link[ComplexHeatmap]{pheatmap}}}

\item{returnValues}{Logical; whether to return the matrix of requested 
values instead of plotting.}

\item{color}{Heatmap colorscale}

\item{number_color}{Values color}
}
\value{
A `Heatmap` showing the overlap coefficient as colors, and the 
  overlap size as values. Alternatively, if `returnValues=TRUE`, a matrix 
  of the requested values.
}
\description{
A wrapper for visualizing pairwise-wise overlaps across multiple sets of
genomic ranges.
}
\examples{
# random list of GRanges:
grl <- lapply(c(A=10,B=20,C=30), FUN=function(x){
  GRanges("seq1", IRanges(runif(x,1,1000), width=20))
})
regionOverlaps(grl)
}
