% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlaps.R
\name{regionsToUpset}
\alias{regionsToUpset}
\title{regionsToUpset}
\usage{
regionsToUpset(
  x,
  reference = c("reduce", "disjoin"),
  returnList = FALSE,
  ignore.strand = FALSE,
  maxgap = -1L,
  minoverlap = 0L,
  ...
)
}
\arguments{
\item{x}{A named list of genomic ranges (or paths to bed files)}

\item{reference}{The method for creating the reference windows ('reduce' or
'disjoin'). Alternatively, a `GRanges` object of reference windows.}

\item{returnList}{Logical; whether to return the list of regions instead of 
plotting.}

\item{ignore.strand}{Logical; whether to ignore strand for overlaps (default
FALSE).}

\item{maxgap}{Max gap between regions to consider an overlap.}

\item{minoverlap}{Minimum number of overlapping bases.}

\item{...}{Further arguments specifying how the overlaps are done, 
passed to \code{\link[GenomicRanges]{findOverlaps-methods}}).}
}
\value{
A data.frame of set inclusions which can be directly input to
  \code{\link[ComplexHeatmap]{make_comb_mat}}, and then
  \code{\link[ComplexHeatmap]{UpSet}}.
}
\description{
Prepares sets of regions for UpSet overlap representation.
}
\examples{
# random list of GRanges:
grl <- lapply(c(A=10,B=20,C=30), FUN=function(x){
  GRanges("seq1", IRanges(runif(x,1,1000), width=20))
})
input_for_upset <- regionsToUpset(grl)
# we would then plot the data with:
ComplexHeatmap::UpSet(make_comb_mat(input_for_upset))
}
