% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{renormalizeBorders}
\alias{renormalizeBorders}
\title{renormalizeBorders}
\usage{
renormalizeBorders(ml, trim = NULL, assay = "input", nWindows = NULL)
}
\arguments{
\item{ml}{A named matrix list as produced by \code{\link{signal2Matrix}}.}

\item{trim}{Quantiles trimmed at each extreme (for linear normalization)}

\item{assay}{Assay to use (ignored unless `ml` is an ESE object), defaults to
the first assay.}
}
\value{
A renormalized list of signal matrices.
}
\description{
This function is \strong{deprecated} and will be replaced by
\code{\link{renormalizeMatrices}}.
Renormalizes a list of signal matrices on the assumption that
the left/right borders of the matrices represent background signal which 
should be equal across samples.
\strong{This is not a safe normalization procedure}: it will work only if 
1) the left/right borders of the matrices are sufficiently far from the 
signal (e.g. peaks), and 2) the signal-to-noise ratio is comparable across 
samples.
}
