% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signalMatrices.R
\name{rescaleSignalMatrices}
\alias{rescaleSignalMatrices}
\title{rescaleSignalMatrices}
\usage{
rescaleSignalMatrices(ml, scaleFactors)
}
\arguments{
\item{ml}{A named matrix list as produced by \code{\link{signal2Matrix}}.}

\item{scaleFactors}{A numeric vector of same length as `ml`, 
indicating the scaling factors by which to multiply each matrix.}
}
\value{
A renormalized list of signal matrices.
}
\description{
rescaleSignalMatrices
}
