% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESE.R
\name{showTrackInfo}
\alias{showTrackInfo}
\title{showTrackInfo}
\usage{
showTrackInfo(x, assay = "input", doPrint = TRUE)
}
\arguments{
\item{x}{A named list of signal matrices or an EnrichmentSE object as 
produced by \code{\link{signal2Matrix}}}

\item{assay}{The assay to use, defaults to the input assay.}

\item{doPrint}{Logical; whether to print the information.}
}
\value{
An invisible list of captions.
}
\description{
Provide some information about the relative signal ranges of each track.
}
