% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal2Matrix.R
\name{signal2Matrix}
\alias{signal2Matrix}
\title{signal2Matrix}
\usage{
signal2Matrix(
  filepaths,
  regions,
  extend = 2000,
  w = NULL,
  scaledBins = 50L,
  type = c("center", "scaled"),
  binMethod = c("max", "mean", "min"),
  BPPARAM = 1L,
  ret = c("list", "ESE"),
  verbose = TRUE,
  ...
)
}
\arguments{
\item{filepaths}{A named vector of filepaths (e.g. to bigwig files; bam files
are also supported, but with limited functionalities). Can also be a named
list including a combination of paths to such files and `GRanges` object.
For `GRanges` objects, the `score` column will be used (absolute coverage
mode).}

\item{regions}{A `GRanges` of the regions/positions around which to plot, or
the path to a bed file of such regions.}

\item{extend}{Number of basepair to extend on either side of the regions. 
Must be a multiple of `w`. Can also be an integer of length 2, indicating 
the extension upstream and downstream.}

\item{w}{Bin size}

\item{scaledBins}{The number of bins for the scale region (ignored if 
`type="center"`)}

\item{type}{Either 'center' (plots fixed-size region around the centers of 
`regions`) or 'scaled' (scales the signal in `regions` and plot 
surroundings)}

\item{binMethod}{Whether to compute the 'max' (default), 'mean' or 'min' per
bin.}

\item{BPPARAM}{A \code{\link[BiocParallel]{BiocParallelParam}} object, or 
the number of threads to use to read and prepare the data. Note that the 
  rate-limiting process is reading from disk, so unless you have an unusually
  fast disk, using multi-threading is actually likely to slow down rather 
  than speed up the process.}

\item{verbose}{Logical; whether to print processing information}

\item{...}{Passed to \code{\link[EnrichedHeatmap]{as.normalizedMatrix}} when
reading bigwig files, or to \code{\link{getBinSignalFromBam}} when reading
bam files.}
}
\value{
A list of `normalizeToMatrix` objects
}
\description{
Reads the signals around (the centers of) a set of regions.
}
\examples{
# we fetch the path to the example bigwig file:
(bw <- system.file("extdata/example_atac.bw", package="epiwraps"))
# we load example regions:
regions <- rtracklayer::import(system.file("extdata/example_peaks.bed", 
                                           package="epiwraps"))
length(regions)
# we obtain the matrix of the signal around the regions:
m <- signal2Matrix(bw, regions)
dim(m[[1]])
# we can plot it with:
plotEnrichedHeatmaps(m)
# we could also take a broader range around the center of the regions, and 
# use bigger bins:
m <- signal2Matrix(bw, regions, extend=2000, w=20)
# the matrix has the same size, but shows broader regions:
dim(m[[1]])
plotEnrichedHeatmaps(m)
}
