% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleRegion.R
\name{signalsAcrossSamples}
\alias{signalsAcrossSamples}
\title{signalsAcrossSamples}
\usage{
signalsAcrossSamples(files, region, ignore.strand = TRUE)
}
\arguments{
\item{files}{A named list of paths to biwgig files or of `GRanges` objects
with a `score` column.}

\item{region}{The region of interest, either given as a string (in the 
"chr:start-end" format) or as a `GRanges` of length 1.}

\item{ignore.strand}{Logical; whether to merge scores from the two strands
given stranded objects.}
}
\value{
A disjoined `GRanges object` with the scores as metadata columns.
}
\description{
Obtain a matrix of score/coverages across a region for a list of BigWig files
or GRanges.
}
