% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bam2bw.R
\name{tileRle}
\alias{tileRle}
\title{tileRle}
\usage{
tileRle(x, bs = 10L, method = c("max", "min", "mean"), roundSummary = FALSE)
}
\arguments{
\item{x}{A numeric `Rle` (or `RleList`)}

\item{bs}{A positive integer specifying the bin size}

\item{method}{The method for summarizing bins}

\item{roundSummary}{Logical; whether to round bins with summarized coverage
(default FALSE)}
}
\value{
An object of same class and length as `x`
}
\description{
Creates an Rle of fixed-with bins from a continuous numeric Rle
}
\examples{
# creating a dummy coverage and visualizing it:
cov <- Rle(rpois(100,0.5))
plot(cov, type="l", col="lightgrey")
# summarizing to tiles of width 5 (by default using maximum)
cov2 <- tileRle(cov, bs=5L)
lines(cov2, col="red")
}
