% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{views2Matrix}
\alias{views2Matrix}
\title{views2Matrix}
\usage{
views2Matrix(v, padVal = NA_integer_)
}
\arguments{
\item{v}{A `RleViews` or `RleViewsList` object with views of the same width.}

\item{padVal}{The value to assign to out-of-bound regions.}
}
\value{
A numeric matrix.
}
\description{
converts a RleViews or RleViewsList with views of the same width to a matrix,
setting out-of-bounds regions to NA (or `padVal`).
}
\examples{
# we create an example RleViews with out-of-bound regions:
library(IRanges)
co <- Rle(values=c(0,1,0), lengths=c(100,50,20))
v <- Views(co, c(25,150),c(50,175))
# convert to matrix
views2Matrix(v)
}
