% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callPeaks.R
\name{callPeaks}
\alias{callPeaks}
\title{callPeaks}
\usage{
callPeaks(
  signalBam,
  controlBam = NULL,
  broad = FALSE,
  fragSize = 125L,
  wWidth = NULL,
  spacing = NULL,
  localSize = NULL,
  minLFC = log2(3),
  prior.count = 3L,
  filter = 1L + prior.count,
  p.threshold = 10^-4,
  blacklist = NULL,
  BPPARAM = NULL,
  verbose = TRUE
)
}
\arguments{
\item{signalBam}{Path to the ChIP bam file}

\item{controlBam}{Optional path to the input (i.e. negative control) bam file}

\item{broad}{Logical, whether to call broad regions rather than narrow peaks}

\item{fragSize}{Fragment length}

\item{wWidth}{Window width. Defaults to twice the fragment length 
(`fragSize`) for narrow peaks and four times for broad.}

\item{spacing}{Shift between sliding windows}

\item{localSize}{Local neighborhood size}

\item{minLFC}{Minimum log2-foldchange}

\item{prior.count}{Prior count added to compute foldchange}

\item{filter}{The minimum count for a window to be considered}

\item{p.threshold}{P-value threshold (poisson probability)}

\item{blacklist}{An optional GRanges of regions to be ignored.}

\item{BPPARAM}{BiocParallel BPPARAM}

\item{verbose}{Logical; whether to print progress messages}
}
\value{
A GRanges object
}
\description{
This is a very simple peak calling function wrapping around several 
\code{\link{csaw}} functions.
}
\details{
The function supports broad and narrow peak calling, as well as calls based
on the comparison to a control (based on csaw's 
\code{\link[csaw]{filterWindowsControl}}), based on a local background for 
narrow peaks (see \code{\link[csaw]{filterWindowsLocal}}), or based on a 
global background for broad peaks (see 
\code{\link[csaw]{filterWindowsGlobal}}).
Note that the caller hasn't been thoroughly tested, and given its 
simplicity it's very likely that it is less robust than more established
methods.
}
