% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callPeaks.R
\name{callPeaks}
\alias{callPeaks}
\title{callPeaks}
\usage{
callPeaks(
  bam,
  ctrl = NULL,
  paired = FALSE,
  type = c("narrow", "broad"),
  blacklist = NULL,
  binSize = 5L,
  fragLength = NULL,
  minPeakCount = 5L,
  minFoldChange = 1.3,
  pthres = 10^-3,
  priorFragLength = 150L,
  bgWindow = c(1, 5, 10) * 1000,
  pseudoCount = 1L,
  outFormat = c("custom", "narrowPeak"),
  verbose = TRUE,
  ...
)
}
\arguments{
\item{bam}{A signal bam file}

\item{ctrl}{An optional (but highly recommended) control bam file}

\item{paired}{Logical, whether the reads are paired}

\item{type}{The type of peaks to identify ('narrow' or 'broad').}

\item{blacklist}{An optional `GRanges` of regions to be excluded (or the path
to such a file). Since the blacklisted regions are removed from both the
signal and control peaks, this also has an important impact on the 
empirical FDR (when `ctrl` is given).}

\item{binSize}{Binsize used to estimate peak shift}

\item{fragLength}{Fragment length. Ignored if `paired=TRUE`. If not given,
will be estimated from the data.}

\item{minPeakCount}{The minimum summit count for a region to be considered.
Decreasing this can substantially increase the running time.}

\item{minFoldChange}{The minimum fold-change for a region to be considered.
Decreasing this can substantially increase the running time.}

\item{pthres}{The p-value threshold to use}

\item{priorFragLength}{Prior fragment length (ignored if `paired=TRUE`). This
is used for coverage computation to estimate peak shift.}

\item{bgWindow}{The windows to consider (in addition to the peak itself) 
for local background.}

\item{outFormat}{The output format ('custom' or 'narrowPeak')}

\item{verbose}{Logical; whether to output progress messages}

\item{...}{Passed to `estimateFragSize`}
}
\value{
A `GRanges`
}
\description{
This is a R-based implementation of the general MACS2 strategy (Zhang et al.,
 Genome Biology 2008), taking some freedom here and there in comparison to 
 the original.
}
