% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callPeaks.R
\name{exportNarrowPeaks}
\alias{exportNarrowPeaks}
\title{exportNarrowPeaks}
\usage{
exportNarrowPeaks(p, file)
}
\arguments{
\item{p}{A GRanges, as produced by `callPeaksExperimental`}

\item{file}{The path to the file where to save}
}
\value{
Invisible file path.
}
\description{
exportNarrowPeaks
}
\examples{
# call some peaks:
bam <- system.file("extdata", "ex1.bam", package="Rsamtools")
peaks <- callPeaksExperimental(bam, paired=TRUE)
# save them:
filepath <- tempfile(fileext="narrowPeak")
exportNarrowPeaks(peaks, filepath)
}
