% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{getCovStats}
\alias{getCovStats}
\title{getCovStats}
\usage{
getCovStats(
  x,
  binSize = 1000,
  nbBins = 10000,
  exclude = NULL,
  genome = NULL,
  BPPARAM = SerialParam()
)
}
\arguments{
\item{x}{A (named) vector of paths to bigwig files}

\item{binSize}{The size of bins}

\item{nbBins}{The number of random bins. More bins gives more accurate 
readouts but take longer to compute.}

\item{exclude}{Region to exclude}

\item{genome}{A named vector of chromosome sizes. If a GRanges without 
seqlengths, the maximum coordinate of each chromosome will be used.}

\item{BPPARAM}{BioParallel BPPARAM for multithreading across files.}
}
\value{
A named list of QC tables
}
\description{
Assembles read distribution statistics from a set of bigwig files based on
random windows.
}
