% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{getNormFactors}
\alias{getNormFactors}
\alias{bwNormFactors}
\title{getNormFactors}
\usage{
getNormFactors(
  x,
  wsize = 10L,
  nwind = 20000L,
  peaks = NULL,
  trim = 0.01,
  useSeqLevels = NULL,
  paired = NULL,
  ...,
  verbose = TRUE,
  method = c("background", "SES", "MAnorm", "S3norm", "2cLinear")
)

bwNormFactors(x, ...)
}
\arguments{
\item{x}{A vector of paths to bigwig files, to bam files, or alternatively a 
list of coverages in RleList format. (Mixed formats are not supported)}

\item{wsize}{The size of the random windows. If any of the bigwig files 
records point events (e.g. insertions) at high resolution (e.g. nucleotide),
use a lot (e.g. >10k) of small windows (e.g. `wsize=10`), as per default
settings. Otherwise the process can be lightened by using fewer bigger 
windows.}

\item{nwind}{The number of random windows}

\item{peaks}{A list of peaks (GRanges) for each experiment in `x`, or a
vector of paths to such files, or a single GRanges of peaks to use for 
MAnorm}

\item{trim}{Amount of trimming when calculating means}

\item{method}{Normalization method (see details)}
}
\value{
A vector of normalization factors, or for the 'S3norm' and '2cLinear'
  methods, a numeric matrix with a number of rows equal to the length 
  of `x`, and two columns indicating the alpha and beta terms.

A vector of normalization factors or, for methods 'S3norm' and 
  '2cLinear', a matrix of per-sample normalization parameters.
}
\description{
Estimates normalization factors for a set of samples (i.e. bam/bigwig files).
}
\details{
The 'background' or 'SES' normalization method (they are synonyms here)
(Diaz et al., Stat Appl Gen et Mol Biol, 2012) assumes that the background
noise should on average be the same across experiments, an assumption that 
works well in practice when there are not very large differences in 
signal-to-noise ratio. The implementation uses the trimmed mean number of 
reads in random windows with non-zero counts.
The 'MAnorm' approach (Shao et al., Genome Biology 
2012) assumes that regions that are commonly enriched (i.e. common peaks) in 
two experiments should on average have the same signal in the two 
experiments. These methods then use linear scaling.
The 'S3norm' (Xiang et al., NAR 2020) and '2cLinear' methods try to 
normalize both simultaneously. S3norm does this in a log-linear fashion (as 
in the publication), while '2cLinear' does it on the original scale.
}
\section{Functions}{
\itemize{
\item \code{bwNormFactors()}: deprecated in favor of getNormFactors

}}
\examples{
# we get an example bigwig file, and use it twice:
bw <- system.file("extdata/example_atac.bw", package="epiwraps")
bw <- c(sample1=bw, sample2=bw)
# we indicate to use only chr1, because it's the only one in the example file
getNormFactors(bw, useSeqLevels="1")
}
