% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESE.R
\name{getSignalMatrices}
\alias{getSignalMatrices}
\title{getSignalMatrices}
\usage{
getSignalMatrices(x, assay = 1L)
}
\arguments{
\item{x}{An object of class `EnrichmentSE`, as produced by 
\code{\link{signal2Matrix}}.}

\item{assay}{The assay to extract (defaults to the first assay).}
}
\value{
A list of normalizedMatrix objects.
}
\description{
Extracts a list of signal matrices from an EnrichmentSE object.
}
