% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signalMatrices.R
\name{meltSignals}
\alias{meltSignals}
\title{meltSignals}
\usage{
meltSignals(ml, fun = c("mean", "sum", "median"))
}
\arguments{
\item{ml}{A named list of matrices as produced by \code{\link{signal2Matrix}}}

\item{fun}{The aggregation to perform. Either "mean", "sum", "median", or
a custom function to be applied on columns.}
}
\value{
A data.frame.
}
\description{
Aggregates and melts a list of signal matrices, for plotting (with ggplot).
}
