% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signalMatrices.R
\name{meltSignals}
\alias{meltSignals}
\title{meltSignals}
\usage{
meltSignals(ml, fun = NULL, splitBy = NULL, trim = 0.98, assay = 1L)
}
\arguments{
\item{ml}{A named list of signal matrices or an EnrichmentSE object as 
produced by \code{\link{signal2Matrix}}}

\item{fun}{An optional custom aggregation function (or named list thereof).}

\item{trim}{The quantile above which to trim values. If a numeric vector of 
length 2, will be used as lower and upper quantiles beyond which to trim.}

\item{assay}{Assay to use (ignored unless `ml` is an ESE object), defaults to
the first assay.}
}
\value{
A data.frame.
}
\description{
Aggregates and melts a list of signal matrices, for plotting (with ggplot).
}
