% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESE.R
\name{ml2ESE}
\alias{ml2ESE}
\title{Creates an EnrichmentSE from a list of normalizedMatrix objects}
\usage{
ml2ESE(ml, assayName = "input", rowRanges = NULL, addScore = FALSE, ...)
}
\arguments{
\item{ml}{A named list of normalizedMatrix objects with corresponding rows.}

\item{assayName}{The name of the assay, defaults to 'input'}

\item{rowRanges}{An optional GRanges object corresponding to the rows of each
object of `ml`.}

\item{addScore}{Logical; whether to add an enriched_score assay.}

\item{...}{Passed to `SummarizedExperiment()`}
}
\description{
Creates an EnrichmentSE from a list of normalizedMatrix objects
}
