% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifs.R
\name{motifCoOccurence}
\alias{motifCoOccurence}
\title{motifCoOccurence}
\usage{
motifCoOccurence(
  motifs,
  pairs,
  regions,
  genome,
  centerDist = TRUE,
  minDist = 5,
  maxDist = 50
)
}
\arguments{
\item{motifs}{A list of motifs (only those specified in `pairs` will be 
used).}

\item{pairs}{A list of pairs of motifs for which to compute co-occurence.
Specifically, this should be a (optionally named) list of character vectors
of length 2, corresponding to names in `motifs`.}

\item{regions}{The regions in which to search for motif matches.}

\item{genome}{A genome object or path to a genome fasta file.}

\item{centerDist}{Logical; whether to compute distances from the center of 
the motifs.}

\item{minDist}{The minimum distance between matches for a co-occurrence. This
is primarily used to exclude interactions between highly-similar, 
overlapping motifs, and to focus on putative cooperative interactions 
between TFs. If interested in competing TFs, set this to <=0.}

\item{maxDist}{The maximum distance between the matches for a co-occurence.}
}
\value{
A list of sparse logical matrices, with one matrix for each value 
  of `minDist`/`maxDist`.
}
\description{
Returns regions that have matches for given pairs of motifs within certain 
distances of each other.
}
\details{
Note that both `minDist` and `maxDist`, rather than specifying a single 
threshold, can compute co-occurence for multiple thresholds (which is must
faster than running the function multiple times). `minDist` and `maxDist` 
should have the same length, and the corresponding entries will be used 
together to produce multiple matrices.
Also note that all matches are stored in memory, so using this function 
across the entire genome is not advisable (unless for very few motifs).
}
