% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCounts.R
\name{peakPbCountsSE}
\alias{peakPbCountsSE}
\title{peakPbCountsSE}
\usage{
peakPbCountsSE(
  fragfile,
  peaks,
  bcmap,
  insertions = FALSE,
  genome = NULL,
  minFragLength = 1L,
  maxFragLength = 5000L
)
}
\arguments{
\item{fragfile}{The path to a Tabix-indexed fragment file.}

\item{peaks}{A GRanges of the regions in which to count.}

\item{bcmap}{A named vector, indicating the pseudobulk sample (values) in 
which to include each barcode (names).}

\item{insertions}{If TRUE, (shifted) Tn5 insertions events are counted 
instead of fragments. This means that each fragment gets counted twice
(for both ends). Default FALSE.}

\item{genome}{A optional genome object or path to a genom fasta file. If
included, GC bias will be added to the rowData of the output object.}

\item{minFragLength}{Minimum fragment length for a fragment to be counted.}

\item{maxFragLength}{Maximum fragment length for a fragment to be counted.}
}
\value{
A \link[SummarizedExperiment]{RangedSummarizedExperiment} with a 
  'counts' assay, and columns corresponding to each unique value of `bcmap`.
}
\description{
Generate a pseudobulk peak counts SummarizedExperiment from a fragment file.
}
