% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signalMatrices.R
\name{plotEnrichedHeatmaps}
\alias{plotEnrichedHeatmaps}
\title{plotEnrichedHeatmaps}
\usage{
plotEnrichedHeatmaps(
  ml,
  trim = 0.998,
  colors = inferno(100),
  scale_title = "density",
  title_size = 11,
  use_raster = NULL,
  row_order = NULL,
  cluster_rows = FALSE,
  ...
)
}
\arguments{
\item{ml}{A named list of matrices as produced by \code{\link{signal2Matrix}}}

\item{trim}{The quantile above which to trim values for the colorscale}

\item{colors}{The heatmap colors to use}

\item{scale_title}{The title of the scale}

\item{title_size}{The size of the heatmap titles}

\item{use_raster}{Passed to EnrichedHeatmap::EnrichedHeatmap}

\item{row_order}{Option order of the rows}

\item{cluster_rows}{Whether to cluster rows}

\item{...}{Passed to EnrichedHeatmap::EnrichedHeatmap}
}
\value{
A HeatmapList object
}
\description{
Plots enrichment heatmaps from the output of `signal2Matrix`
}
\examples{
# we first fetch the path to the example bigwig file:
bw <- system.file("extdata/example_atac.bw", package="epiwraps")
# Since we only have one, we'll use the same and pretend they're 2 samples:
bw <- c(sample1=bw, sample2=bw)
# we next load regions of interest (either GRanges or path to a bed file):
regions <- system.file("extdata/example_peaks.bed", package="epiwraps")
# we obtain the matrix of the signal around the regions:
m <- signal2Matrix(bw, regions)
plotEnrichedHeatmaps(m)
# we could also just plot one with:
# plotEnrichedHeatmaps(m[1])
# or change the aesthetics, e.g.:
plotEnrichedHeatmaps(m, trim=0.98, scale_title="RPKM", 
                     colors=c("white","darkred"))
# any argument accepted by `EnrichedHeatmap` (and hence by 
# `ComplexHeatmap::Heatmap`) can be used, e.g.: 
plotEnrichedHeatmaps(m, row_title="My regions of interest")
}
