% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signalMatrices.R
\name{plotEnrichedHeatmaps}
\alias{plotEnrichedHeatmaps}
\title{plotEnrichedHeatmaps}
\usage{
plotEnrichedHeatmaps(
  ml,
  trim = 0.998,
  colors = viridisLite::inferno(100),
  scale_title = "density",
  title_size = 11,
  row_order = NULL,
  cluster_rows = FALSE,
  ...
)
}
\arguments{
\item{ml}{A named list of matrices as produced by `signal2Matrix`}

\item{trim}{The quantile above which to trim values for the colorscale}

\item{colors}{The heatmap colors to use}

\item{scale_title}{The title of the scale}

\item{title_size}{The size of the heatmap titles}

\item{row_order}{Option order of the rows}

\item{cluster_rows}{Whether to cluster rows}

\item{...}{Passed to EnrichedHeatmap::EnrichedHeatmap}
}
\value{
A HeatmapList object
}
\description{
Plots enrichment heatmaps from the output of `signal2Matrix`
}
