% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlaps.R
\name{regionCAT}
\alias{regionCAT}
\title{regionCAT}
\usage{
regionCAT(
  regions1,
  regions2,
  start = 5L,
  concord.type = c("both", "inTop", "inAll"),
  returnData = FALSE,
  ignore.strand = TRUE
)
}
\arguments{
\item{regions1}{A GRanges object}

\item{regions2}{A GRanges object}

\item{start}{The rank at which to start plotting (removes large variations
at the beginning when very few regions are considered)}

\item{concord.type}{Concordance type to plot, either 'inTop', 'inAll', or 
'both' (see details). Ignored if `returnData=TRUE`.}

\item{returnData}{Logical; whether to return the data instead of plotting.}

\item{ignore.strand}{Logical; whether to ignore the strand for computing
overlap (default TRUE)}
}
\value{
A ggplot object, or a data.frame if `returnData=TRUE`.
}
\description{
Computes/plots the 'concordance at the top' (CAT) of two lists of genomic
  regions.
}
\details{
The two concordance types are as follows:
* 'inTop' indicates the proportion of the top X regions that are in the top 
  X in both lists.
* 'all' indicates the proportion of the top X regions that are anywhere in 
  the other list (since this relationship is asymmetrical, the mean of both
  two directions is used).
}
