% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlaps.R
\name{regionOverlaps}
\alias{regionOverlaps}
\title{regionOverlaps}
\usage{
regionOverlaps(
  listOfRegions,
  ignore.strand = TRUE,
  cluster = length(listOfRegions) > 2,
  color = viridis::plasma(100),
  number_color = "black",
  ...
)
}
\arguments{
\item{listOfRegions}{A named list of two or more (non-empty) `GRanges`}

\item{ignore.strand}{Logical; whether to ignore strand for overlaps}

\item{cluster}{Logical; whether to cluster rows/columns}

\item{color}{Heatmap colorscale}

\item{number_color}{Values color}

\item{...}{Passed to \code{\link[ComplexHeatmap]{pheatmap}}}
}
\value{
A `Heatmap` showing the overlap coefficient as colors, and the 
  overlap size as values.
}
\description{
A wrapper for visualizing pairwise-wise overlaps across multiple sets of
genomic ranges.
}
\examples{
# random list of GRanges:
grl <- lapply(c(A=10,B=20,C=30), FUN=function(x){
  GRanges("seq1", IRanges(runif(x,1,1000), width=20))
})
regionOverlaps(grl)
}
