% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlaps.R
\name{regionOverlaps}
\alias{regionOverlaps}
\title{regionOverlaps}
\usage{
regionOverlaps(
  listOfRegions,
  ignore.strand = TRUE,
  color = viridis::plasma(100),
  number_color = "black",
  ...
)
}
\arguments{
\item{listOfRegions}{A named list of two or more (non-empty) `GRanges`}

\item{ignore.strand}{Logical; whether to ignore strand for overlaps}

\item{color}{Heatmap colorscale}

\item{number_color}{Values color}

\item{...}{Passed to \code{\link[ComplexHeatmap]{pheatmap}}}
}
\value{
A `Heatmap` showing the overlap coefficient as colors, and the 
  overlap size as values.
}
\description{
regionOverlaps
}
