% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signalMatrices.R
\name{renormalizeBorders}
\alias{renormalizeBorders}
\title{renormalizeBorders}
\usage{
renormalizeBorders(
  ml,
  method = "linear",
  trim = NULL,
  nWindows = max(floor(ncol(ml[[1]])/10), 1)
)
}
\arguments{
\item{ml}{A named matrix list as produced by \code{\link{signal2Matrix}}.}

\item{method}{Either "linear", or a normalization method, passed to 
\code{\link[edgeR]{calcNormFactors}}.}

\item{trim}{Quantiles trimmed at each extreme (for linear normalization)}
}
\value{
A renormalized list of signal matrices.
}
\description{
This function renormalizes a list of signal matrices on the assumption that
the left/right borders of the matrices represent background signal which 
should be equal across samples.
\strong{This is not a safe normalization procedure}: it will work only if 
1) the left/right borders of the matrices are sufficiently far from the 
signal (e.g. peaks), and 2) the signal-to-noise ratio is comparable across 
samples.
}
