% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signalMatrices.R
\name{signal2Matrix}
\alias{signal2Matrix}
\title{signal2Matrix}
\usage{
signal2Matrix(
  filepaths,
  regions,
  extend = 1000,
  w = 10,
  cuts = FALSE,
  ...,
  RPM = TRUE,
  BPPARAM = BiocParallel::SerialParam(),
  flgs = Rsamtools::scanBamFlag(isDuplicate = FALSE, isSecondaryAlignment = FALSE)
)
}
\arguments{
\item{filepaths}{A named vector of filepaths (preferentially to bigwig files)}

\item{regions}{A `GRanges` of the regions/positions around which to plot}

\item{extend}{Number of basepair to extend on either side of the regions}

\item{w}{Bin size}

\item{cuts}{Whether to count cuts (e.g. beginning/end of fragments) rather
than coverage (ignored unless the input are bam files)}

\item{...}{Passed to `EnrichedHeatmap::normalizeToMatrix`}

\item{RPM}{Whether to perform RPM normalization (for bam input)}

\item{BPPARAM}{Not yet implemented}

\item{flgs}{Flags for bam reading}
}
\value{
A list of `normalizeToMatrix` objects
}
\description{
Reads the signals around a set of regions.
}
