% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signalMatrices.R
\name{signal2Matrix}
\alias{signal2Matrix}
\title{signal2Matrix}
\usage{
signal2Matrix(
  filepaths,
  regions,
  extend = 1000,
  w = 10,
  cuts = FALSE,
  ...,
  RPM = TRUE,
  BPPARAM = 1L,
  flgs = scanBamFlag(isDuplicate = FALSE, isSecondaryAlignment = FALSE)
)
}
\arguments{
\item{filepaths}{A named vector of filepaths (preferentially to bigwig files)}

\item{regions}{A `GRanges` of the regions/positions around which to plot, or
the path to a bed file of such regions.}

\item{extend}{Number of basepair to extend on either side of the regions}

\item{w}{Bin size}

\item{cuts}{Whether to count cuts (e.g. beginning/end of fragments) rather
than coverage (ignored unless the input are bam files)}

\item{...}{Passed to \code{\link[EnrichedHeatmap]{normalizeToMatrix}}.}

\item{RPM}{Whether to perform RPM normalization (for bam input)}

\item{BPPARAM}{A \code{\link[BiocParallel]{BiocParallelParam}} object, or 
the number of threads to use to read and prepare the data.}

\item{flgs}{Flags for bam reading}
}
\value{
A list of `normalizeToMatrix` objects
}
\description{
Reads the signals around (the centers of) a set of regions.
}
\examples{
# we fetch the path to the example bigwig file:
(bw <- system.file("extdata/example_atac.bw", package="epiwraps"))
# we load example regions:
regions <- rtracklayer::import(system.file("extdata/example_peaks.bed", 
                                           package="epiwraps"))
length(regions)
# we obtain the matrix of the signal around the regions:
m <- signal2Matrix(bw, regions)
dim(m[[1]])
# we can plot it with:
plotEnrichedHeatmaps(m)
# we could also take a broader range around the center of the regions, and 
# use bigger bins:
m <- signal2Matrix(bw, regions, extend=2000, w=20)
# the matrix has the same size, but shows broader regions:
dim(m[[1]])
plotEnrichedHeatmaps(m)
}
