% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabixChrChunkApply.R
\name{tabixChrApply}
\alias{tabixChrApply}
\title{tabixChrApply}
\usage{
tabixChrApply(
  x,
  fn,
  keepSeqLvls = NULL,
  exclude = NULL,
  only = NULL,
  BPPARAM = SerialParam(),
  ...
)
}
\arguments{
\item{x}{The path to a tabix-indexed bam file, or a TabixFile object.}

\item{fn}{The function to be run, the first argument of which should be a
`GRanges`}

\item{keepSeqLvls}{An optional vector of seqLevels to keep}

\item{only}{}

\item{BPPARAM}{A `BiocParallel` parameter object for multithreading. Note 
that if used, memory usage will be high; in this context we recommend a 
high `nChunks`.}

\item{...}{Passed to `fn`}
}
\value{
A list of whatever `fn` returns
}
\description{
Runs a function on reads/fragments from each chromosomes of a Tabix-indexed 
fragment file. This is especially used by other functions to 
avoid loading all alignments into memory, or to parallelize reads processing.
}
